+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v20.0.1...HEAD)
+### Changed
+* [STREAM-1784](https://inindca.atlassian.net/browse/STREAM-1784) - Properly handle session not found for terminate stanza
 
 # [v20.0.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v20.0.0...v20.0.1)
 ### Changed